Добавлено изменение цвета для категории
Пока это влияет только на графики
