+++ app/models/category.rb
@@ -2,4 +2,5 @@ class Category < ActiveRecord::Base
   has_many :operations
 
   validates :title, :presence => true
+  validates :color, :uniqueness => true
 end
