+++ app/views/categories/_form.html.erb
@@ -9,13 +9,23 @@
           </ul>
         </div>
     <% end %>
-    <div class="form-group">
-      <%= f.label :title, 'Название категории', :class => 'col-sm-2 control-label', :autofocus => true %>
-      <div class="col-sm-10">
-        <%= f.text_field :title, :class => 'form-control', :placeholder => 'Название' %>
-      </div>
+    <div class="col-md-6">
+        <div class="form-group">
+            <%= f.label :title, 'Название категории', :class => 'col-sm-2 control-label', :autofocus => true %>
+            <div class="col-sm-10">
+                <%= f.text_field :title, :class => 'form-control', :placeholder => 'Название' %>
+            </div>
+        </div>
+    </div>
+    <div class="col-md-6">
+        <div class="form-group">
+            <%= f.label :color, 'Цвет', :class => 'col-sm-2 control-label', :autofocus => true %>
+            <div class="col-sm-10">
+                <%= f.color_field :color, :class => 'form-control', :placeholder => 'Цвет' %>
+            </div>
+        </div>
     </div>
     <div class="action">
-      <%= f.submit 'Сохранить', :class => 'btn btn-primary', data: {disable_with: "Сохраняем..."} %>
+        <%= f.submit 'Сохранить', :class => 'btn btn-primary', data: {disable_with: "Сохраняем..."} %>
     </div>
 <% end %>
