+++ app/views/operations/_form.html.erb
@@ -54,7 +54,7 @@
             <div class="form-group">
                 <%= f.label :category_id, 'Категория', :class => 'col-sm-2 control-label' %>
                 <div class="col-sm-10">
-                    <%= select2_collection('operation', 'category_id', Category.where('parent_id is null'), 'id', 'title', @operation.category_id, {:class => 'form-control'}) %>
+                    <%= select2_collection('operation', 'category_id', Category.all, 'id', 'title', @operation.category_id, {:class => 'form-control'}) %>
                 </div>
             </div>
             <div class="form-group">
