+++ db/migrate/20160104202001_add_column_color_to_categories.rb
@@ -1,0 +1,5 @@
+class AddColumnColorToCategories < ActiveRecord::Migration
+  def change
+    add_column :categories, :color, :string
+  end
+end
